[typescript] Move print code to separate file.
