+++ to.etc.domui/src/main/resources/resources/js/odomui.js
@@ -42,7 +42,6 @@ if(WebUI === undefined)
 	WebUI = new Object();
 
 $.extend(WebUI, {
-
 	returnKeyPress : function(evt, node) {
 		var keyCode = WebUI.normalizeKey(evt);
 		if (keyCode != 13000 && keyCode != 13)
@@ -68,269 +67,9 @@ $.extend(WebUI, {
 	oddChar : function(obj) {
 		WebUI.toClip(obj.innerHTML);
 	},
-	/** ******************* Scheduler component ************************ */
-	/** Dynamically loading stylesheets and javascript files (Header Contributer delta's) **/
-	/**
-	 * Load the specified stylesheet by creating a script tag and inserting it @ head.
-	 */
-	loadStylesheet: function(path) {
-		var head = document.getElementsByTagName("head")[0];
-		if(! head)
-			throw "Headless document!?";
-		var link = document.createElement('link');
-		link.type = 'text/css';
-		link.rel = 'stylesheet';
-		link.href = path;
-		link.media = 'screen';
-		head.appendChild(link);
-	},
-
-	loadJavascript: function(path) {
-		var head = document.getElementsByTagName("head")[0];
-		if(! head)
-			throw "Headless document!?";
-		var scp = document.createElement('script');
-		scp.type = 'text/javascript';
-		scp.src = path;
-		head.appendChild(scp);
-	},
-
-	/** Prevents default action to be executed if IE11 is detected */
-	preventIE11DefaultAction : function(e){
-		if((navigator.userAgent.match(/Trident\/7\./))){
-			e.preventDefault();
-		}
-	},
-
-	/** ***************** File upload stuff. **************** */
-	fileUploadChange : function(e) {
-		var tgt = e.currentTarget || e.srcElement;
-		var vv = tgt.value.toString();
-
-		if (!vv) { //IE when setting the value to empty..
-			return;
-		}
-
-		// -- Check extensions,
-		var allowed = tgt.getAttribute('fuallowed');
-		if (allowed) {
-			var ok = false;
-			var spl = allowed.split(',');
-
-			vv = vv.toLowerCase();
-			for(var i = 0; i < spl.length; i++) {
-				var ext = spl[i].toLowerCase();
-				if(ext.substring(0, 1) != ".")
-					ext = "."+ext;
-				if(ext == ".*") {
-					ok = true;
-					break;
-				} else if(vv.indexOf(ext, vv.length - ext.length) !== -1) {
-					ok = true;
-					break;
-				}
-			}
-
-			if (!ok) {
-				var parts = vv.split('.');
-				alert(WebUI.format(WebUI._T.uploadType, (parts.length > 1) ? parts.pop() : '', allowed));
-				tgt.value = "";
-				return;
-			}
-		}
-
-		if(typeof tgt.files[0] !== "undefined") {
-			var s = tgt.getAttribute('fumaxsize');
-			var maxSize = 0;
-			try {
-				maxSize = Number(s);
-			} catch(x) {
-			}
-			if(maxSize <= 0 || maxSize === NaN)
-				maxSize = 100 * 1024 * 1024;						// Default max size is 100MB
-
-			var size = tgt.files[0].size;
-			if(size > maxSize) {
-				alert(WebUI.format(WebUI._T.buplTooBig, maxSize));
-				tgt.value = "";
-				return;
-			}
-		}
-
-		// -- Step 2: create or locate an iframe to handle the upload;
-		var iframe = document.getElementById('webuiif');
-		if (iframe) {
-			iframe.parentNode.removeChild(iframe);
-			iframe = undefined;
-		}
-		if (!iframe) {
-			if(jQuery.browser.msie && ! WebUI.isNormalIE9plus()) {			// MicroMorons report ie9 for ie7 emulation of course
-				// -- IE's below 8 of course have trouble. What else.
-				iframe = document
-					.createElement('<iframe name="webuiif" id="webuiif" src="#" style="display:none; width:0; height:0; border:none" onload="WebUI.ieUpdateUpload(event)">');
-				document.body.appendChild(iframe);
-			} else {
-				iframe = document.createElement('iframe');
-				iframe.id = 'webuiif';
-				iframe.name = "webuiif";
-				iframe.src = "#";
-				iframe.style.display = "none";
-				iframe.style.width = "0px";
-				iframe.style.height = "0px";
-				iframe.style.border = "none";
-				iframe.onload = function() {
-					WebUI.updateUpload(iframe.contentDocument);
-				};
-				document.body.appendChild(iframe);
-			}
-		}
-
-		// -- replace the input with a busy indicator
-		var form = tgt.parentNode; // Get form to submit to later on,
-
-		var img = document.createElement('img');
-		img.border = "0";
-		img.src = DomUIProgressURL;
-		form.parentNode.insertBefore(img, form);
-		form.style.display = 'none';
-
-		// -- Target the iframe
-		form.target = "webuiif"; // Fake a new thingy,
-		form.submit(); // Force submit of the thingerydoo
-		WebUI.blockUI();									// since 20160226: block UI during upload
-	},
-
-	/**
-	 * Called for ie upload garbage only, this tries to decode the utter devastating mess that
-	 * ie makes from xml uploads into an iframe in ie8+. Sigh. The main problem with IE is that
-	 * the idiots that built it mess up XML documents sent to it: when the iframe we use receives
-	 * an xml document the idiot translates it into an html document- with formatted tags and minus
-	 * signs before it to collapse them. Instead of just /rendering/ that they actually /generate/
-	 * that as the content document of the iframe- so getting that means you get no XML at all.
-	 * This abomination was more or less fixed in ie9 - but of course they fucked up compatibility mode
-	 * badly.
-	 *
-	 * Everything below ie9
-	 * ====================
-	 * Everything below ie9 has a special property "XMLDocument" attached to the "document" property of
-	 * the window. This property contains the original XML that was used as an XML tree. So for these
-	 * browsers we just get that and move on.
-	 *
-	 * IE9 in native mode
-	 * ==================
-	 * IE9 in native mode does not mess up the iframe content document, so this mode takes the path of
-	 * all browsers worth the name. Since the original problem was solved the XMLDocument property no
-	 * longer exists.
-	 *
-	 * IE9 in compatibility mode (IE7)
-	 * ===============================
-	 * This gets funny. In this mode the browser /still/ messes up the iframe's content model like the
-	 * old IE's it emulates. But of course the XMLDocument property is helpfully removed - EVEN IN THIS
-	 * MODE.
-	 * Remember: this mode is also entered if you are part of a frameset or iframe that is part of an
-	 * mode: the topmost frameset/page determines the mode for the entire set of pages - they are that
-	 * stupid.
-	 *
-	 * I know no other "workaround" than the horrible concoction below; please turn away if you have
-	 * a weak stomach....
-	 *
-	 * In this case we get the "innerText" content of the iframe. This differers from innerHTML in that
-	 * all html tags that IE added are removed, and only the text is retained. Since the html was generated
-	 * by IE in such a way that the xml was presented to the user - this is actually most of our XML...
-	 * But it contains some problems:
-	 * - there is extra whitespace around it's edges, so we need to remove that..
-	 * - All tags start on a new line with a '-' sign before them... Remove all that...
-	 * - The result will have the &amp; entity replaced by & because they are really stupid, again. So replace
-	 *   that too.
-	 * The resulting thing can sometimes be parsed as XML and then processing continues. But it is far
-	 * from perfect. The biggest problem is that the resulting xml has not properly reserved the original
-	 * whitespace; this may lead to rendering problems.
-	 *
-	 * But as far as I know no other solution to this stupifying bug is possible. Please prove me wrong.
-	 *
-	 * @param e
-	 */
-	ieUpdateUpload : function(e) { // Piece of crap
-		var iframe = document.getElementById('webuiif');
-		var xml;
-		if(iframe.contentWindow && iframe.contentWindow.document.XMLDocument) {
-			xml = iframe.contentWindow.document.XMLDocument; // IMPORTANT Fucking MS Crap!!!! See http://p2p.wrox.com/topic.asp?whichpage=1&TOPIC_ID=62981&#153594
-		} else if(iframe.contentDocument) {
-			var crap = iframe.contentDocument.body.innerText;
-			crap = crap.replace(/^\s+|\s+$/g, ''); // trim
-			crap = crap.replace(/(\n|\r)-*/g, ''); // remove '\r\n-'. The dash is optional.
-			crap = crap.replace(/&/g, '&amp;');		// Replace & with entity
-//			alert('crap='+crap);
-			if(window.DOMParser) {
-				var parser = new DOMParser();
-				xml = parser.parseFromString(crap);
-			} else if(window.ActiveXObject) {
-				xml = new ActiveXObject("Microsoft.XMLDOM");
-				if(! xml.loadXML(crap)) {
-					alert('ie9 in emulation mode unfixable bug: cannot parse xml');
-					window.location.href = window.location.href;
-					WebUI.unblockUI();
-					return;
-				}
-			} else {
-				alert('No idea how to parse xml today.');
-			}
-		} else
-			alert('IE error: something again changed in xml source structure of the iframe, sigh');
-
-		WebUI.updateUpload(xml, iframe);
-	},
-	updateUpload : function(doc, ifr) {
-		try {
-			jQuery.fn.executeDeltaXML(doc);
-		} catch (x) {
-			alert(x);
-			throw x;
-		} finally {
-			WebUI.unblockUI();
-		}
-		/*
-		 * 20081015 jal De iframe zou verwijderd moeten worden, maar als ik dat
-		 * doe onder Firefox dan blijft de browser in "busy" mode staan, met een
-		 * rode enabled "stop" knop en een "busy" mouse.. Voor nu hergebruiken
-		 * we de bestaande iframe dan maar en accepteren de verloren resources.
-		 */
-		// iframe.parentNode.removeChild(iframe); // Suicide..
-	},
 
 
-	unloaded : function() {
-		WebUI._ignoreErrors = true;
-		WebUI.sendobituary();
-	},
 
-	beforeUnload: function() {
-		//-- Make sure no "ajax" errors are reported.
-		WebUI._ignoreErrors = true;
-	},
-
-	/**
-	 * Called at page unload time, this quickly tries to send an obituary to the
-	 * server. This is currently unconditional but can later be augmented to
-	 * send the obituary only when the browser window closes.
-	 */
-	sendobituary : function() {
-		try {
-			var rq;
-			if (window.XMLHttpRequest) {
-				rq = new XMLHttpRequest();
-			} else if (window.ActiveXObject) {
-				rq = new ActiveXObject("Microsoft.XMLHTTP");
-			} else {
-				alert("Cannot send obituary (no transport)");
-				return;
-			}
-			rq.open("GET", DomUI.getObituaryURL() + "?$cid=" + DomUICID + "&webuia=OBITUARY&$pt=" + DomUIpageTag, false);
-			rq.send(null);
-		} catch(ex) {
-//			alert("Sending obit failed:"+ex);
-		}
-	},
 
 
 	_selectStart : undefined,
@@ -367,10 +106,6 @@ $.extend(WebUI, {
 		if(document.attachEvent){
 			document.attachEvent( "onselectstart", WebUI.preventSelection);
 		}
-	},
-
-	preventSelection : function(){
-		return false;
 	},
 
 	dragMouseUp : function() {
@@ -600,332 +335,6 @@ $.extend(WebUI, {
 
 
 	/** ***************** ScrollableTabPanel stuff. **************** */
-	_ignoreScrollClick: 0,
-
-	scrollLeft : function(bLeft) {
-		if(this._ignoreScrollClick != 0 || $(bLeft).hasClass('ui-stab-dis'))
-			return;
-
-		var scrlNavig = $(bLeft.parentNode);
-		var offset = -1 * parseInt($('ul',scrlNavig).css('marginLeft'));
-		var diff = $(scrlNavig).width() - 2 * $('.ui-stab-scrl-left',scrlNavig).width();
-		var me = this;
-		var disa = false;
-		if ( diff >= offset ){
-			disa = true;
-			diff = offset;
-		}
-		this._ignoreScrollClick++;
-		$('ul',scrlNavig).animate({marginLeft: '+=' + diff}, 400, 'swing', function() {
-			$('.ui-stab-scrl-right', scrlNavig).removeClass('ui-stab-dis');
-			if(disa){
-				$(bLeft).addClass('ui-stab-dis');
-			}
-			me._ignoreScrollClick--;
-		});
-	},
-
-	scrollRight : function(bRight) {
-		if(this._ignoreScrollClick != 0 || $(bRight).hasClass('ui-stab-dis'))
-			return;
-
-		var scrlNavig = $(bRight.parentNode);
-		var tabsTotalWidth = $('li:last',scrlNavig).width() + 8 /* paddong = 8 */ + $('li:last',scrlNavig).offset().left - $('li:first',scrlNavig).offset().left;
-		var tabsVisibleWidth = $(scrlNavig).width() - 2 * $('.ui-stab-scrl-right',scrlNavig).width();
-		var maxLeftOffset = tabsTotalWidth - tabsVisibleWidth;
-		var diff = tabsVisibleWidth;
-		var offset = -1 * parseInt($('ul',scrlNavig).css('marginLeft'));
-
-		var disa = false;
-		if (offset >= maxLeftOffset){
-			return;
-		} else if (diff + offset >= maxLeftOffset){
-			diff = maxLeftOffset - offset;
-			disa = true;
-		}
-		this._ignoreScrollClick++;
-		var me = this;
-		$('ul', scrlNavig ).animate({marginLeft: '-=' + diff},400, 'swing', function() {
-			$('.ui-stab-scrl-left', scrlNavig).removeClass('ui-stab-dis');
-			if (disa){
-				$(bRight).addClass('ui-stab-dis');
-			}
-			me._ignoreScrollClick--;
-		});
-	},
-
-	recalculateScrollers : function(scrlNavigId){
-		var scrlNavig = document.getElementById(scrlNavigId);
-		var tabsTotalWidth = $('li:last',scrlNavig).width() + 8 /* paddong = 8 */ + $('li:last',scrlNavig).offset().left - $('li:first',scrlNavig).offset().left;
-		var tabsVisibleWidth = $(scrlNavig).width() - 2 * $('.ui-stab-scrl-right',scrlNavig).width();
-		//WebUI.debug('debug1', 50, 150, 'width:' + ($('li:last',scrlNavig).width() + 8 + $('li:last',scrlNavig).offset().left - $('li:first',scrlNavig).offset().left));
-		//WebUI.debug('debug2', 50, 200, 'offsetX:' + ($('li:first',scrlNavig).offset().left - $('ul', scrlNavig).offset().left));
-
-		if(tabsTotalWidth > tabsVisibleWidth){
-			var leftM = parseInt($('ul',scrlNavig).css('marginLeft'));
-			//WebUI.debug('debug2', 50, 200, 'leftM:' + leftM);
-			if (tabsTotalWidth + leftM > tabsVisibleWidth){
-				$('.ui-stab-scrl-right',scrlNavig).removeClass('ui-stab-dis');
-			}else{
-				$('.ui-stab-scrl-right',scrlNavig).addClass('ui-stab-dis');
-			}
-			if (leftM < 0){
-				$('.ui-stab-scrl-left',scrlNavig).removeClass('ui-stab-dis');
-			}else{
-				$('.ui-stab-scrl-left',scrlNavig).addClass('ui-stab-dis');
-			}
-		}else{
-
-			// 20121115 btadic: we just want to render scroll buttons without arrows and actions. We don't want empty spaces.
-			//$('.ui-stab-scrl-left',scrlNavig).css('visibility','hidden');
-			//$('.ui-stab-scrl-right',scrlNavig).css('visibility','hidden');
-			$('.ui-stab-scrl-left',scrlNavig).css('display','none');
-			$('.ui-stab-scrl-right',scrlNavig).css('display','none');
-			$('ul', scrlNavig).animate({marginLeft: 0}, 400, 'swing');
-		}
-	},
-
-	/**
-	 * This adds a "resize" listener to the window, and every window "resize" it will call a method
-	 * to recalculate the height of a single div (flexid) based on the position of:
-	 * <ul>
-	 *	<li>The 'bottom' position of the element just above it (get it's top position and add it's height)</li>
-	 *	<li>A bottom 'offset', currently an integer in pixels. It will become a "bottom item" id later, so we can auto-size a div "sandwiched" between two other elements.</li>
-	 * </ul>
-	 * The recalculate code will be called every time the window resizes.
-	 *
-	 * @param topid
-	 * @param flexid
-	 * @param bottom
-	 */
-	autoHeightReset: function(topid, flexid, bottom) {
-		$(window).bind("resize", function() {
-			WebUI.recalculateAutoHeight(topid, flexid, bottom);
-		});
-		WebUI.recalculateAutoHeight(topid, flexid, bottom);
-	},
-
-	recalculateAutoHeight: function(topid, flexid, bottom) {
-		try {
-			var tbot = $(topid).offset().top + $(topid).height();
-			var height = $(window).height() - tbot - bottom;
-			$(flexid).height(height+"px");
-		} catch(x) {
-			//-- Ignore for now
-		}
-	},
-
-	setThreePanelHeight: function(top, middle, bottom) {
-//		try {
-		middle = "#"+middle;
-		var height = $(middle).parent().height();		// Assigned height of the container
-
-		var theight = 0;
-		if(typeof top === "string") {
-			theight = $("#"+top).height();					// Get the end of the "top" element
-		} else if(top) {
-			theight = top;
-		}
-
-		//-- Get height of bottom, if present
-		var bheight = 0;
-		if(typeof bottom === "string") {
-			bheight = $("#"+bottom).height();
-		} else if(bottom) {
-			tbottom = bottom;
-		}
-		height -= theight - bheight;
-		if(height < 0) {
-			height = 0;
-		}
-		$(middle).height(height+"px");
-		$(middle).css({"overflow-y": "auto"});
-//		} catch(x) {
-//			//-- Ignore for now
-//		}
-
-	},
-
-	autoHeight: function(flexid, bottom) {
-		$(window).bind("resize", function() {
-			WebUI.autoHeightRecalc(flexid, bottom);
-		});
-		WebUI.autoHeightRecalc(flexid, bottom);
-	},
-
-	autoHeightRecalc: function(flexid, bottom) {
-		var tbot = $(flexid).offset().top;
-		var height = $(window).height() - tbot - bottom;
-		$(flexid).height(height + "px");
-	},
-
-	notifySizePositionChangedOnId: function(elemId) {
-		var element = document.getElementById(elemId);
-		if (!element){
-			return;
-		}
-		// Send back size information to server
-		var fields = new Object();
-		fields.webuia = "notifyClientPositionAndSize";
-		fields[element.id + "_rect"] = $(element).position().left + "," + $(element).position().top + "," + $(element).width() + "," + $(element).height();
-		fields["window_size"] = window.innerWidth + "," + window.innerHeight;
-		WebUI.scall(element.id, "notifyClientPositionAndSize", fields);
-	},
-
-	notifySizePositionChanged: function(event, ui) {
-		var element = ui.helper.get(0);
-		if (!element){
-			return;
-		}
-		WebUI.notifySizePositionChangedOnId(element.id);
-	},
-
-	/** *************** Debug thingy - it can be used internaly for debuging javascript ;) ************** */
-	debug : function(debugId, posX, posY, debugInfoHtml) {
-		//Be aware that debugId must not start with digit when using FF! Just lost 1 hour to learn this...
-		if ("0123456789".indexOf(debugId.charAt(0)) > -1){
-			alert("debugId(" + debugId+ ") starts with digit! Please use different one!");
-		}
-		var debugPanel = document.getElementById(debugId);
-		if (null == debugPanel){
-			debugPanel = document.createElement(debugId);
-			$(debugPanel).attr('id', debugId);
-			$(debugPanel).css('position', 'absolute');
-			$(debugPanel).css('marginLeft', 0);
-			$(debugPanel).css('marginTop', 0);
-			$(debugPanel).css('background-color', 'yellow');
-			$(debugPanel).css('border', '1px');
-			$(debugPanel).css('z-index', 2000);
-			$(debugPanel).appendTo('body');
-		}
-		$(debugPanel).css('left', posX);
-		$(debugPanel).css('top', posY);
-		$(debugPanel).html(debugInfoHtml);
-	},
-
-	/*-- Printing support --*/
-	_frmIdCounter: 0,
-
-	backgroundPrint : function(url) {
-		try {
-			var frmname = WebUI.priparePrintDiv(url);
-			WebUI.framePrint(frmname);
-		} catch (x) {
-			alert("Failed: " + x);
-		}
-	},
-
-	framePrint : function(frmname){
-		if (jQuery.browser.msie) {
-			WebUI.documentPrintIE(frmname);
-		} else {
-			WebUI.documentPrintNonIE(frmname);
-		}
-	},
-
-	documentPrintIE : function(frmname) {
-		try {
-			var frm = window.frames[frmname];
-			$("#"+frmname).load(function() {
-				try {
-					frm.focus();
-					setTimeout(function() {
-						if (!frm.document.execCommand('print', true, null)){
-							alert('cannot print: '+x);
-						}
-					}, 1000);
-				} catch(x) {
-					alert('cannot print: '+x);
-				}
-			});
-		} catch(x) {
-			alert("Failed: "+x);
-		}
-	},
-
-	documentPrintNonIE : function(frmname) {
-		try {
-			var frm = window.frames[frmname];
-			$("#"+frmname).load(function() {
-				try {
-					frm.focus();
-					setTimeout(function() {
-						frm.print();
-					}, 1000);
-				} catch(x) {
-					alert('cannot print: '+x);
-				}
-			});
-		} catch(x) {
-			alert("Failed: "+x);
-		}
-	},
-
-	priparePrintDiv : function(url){
-		// Create embedded sizeless div to contain the iframe, invisibly.
-		var div = document.getElementById('domuiprif');
-		if (div)
-			div.innerHTML = "";
-		else {
-			div = document.createElement('div');
-			div.id = 'domuiprif';
-			div.className = 'ui-printdiv';
-			document.body.appendChild(div);
-		}
-
-		// -- Create an iframe loading the required thingy.
-		var frmname = "dmuifrm" + (WebUI._frmIdCounter++); // Create unique
-		// name to
-		// circumvent
-		// ffox "print
-		// only once"
-		// bug
-		if (url.trim() !== "") {
-			$(div).html('<iframe id="' + frmname + '" name="' + frmname + '" src="' + url + '">');
-		} else {
-			//well, this is simple element printing, so we have some size limitations
-			$(div).html('<iframe id="' + frmname + '" name="' + frmname + '" width="1000px" height="1000px"/>');
-		}
-		return frmname;
-	},
-
-	/*-- Printing support for simple text messages. Parameter is id of input/textarea tag that contrains text to be printed out. --*/
-	printtext: function (id) {
-		var item = document.getElementById(id);
-		var textData;
-		if(item && (item.tagName == "input" || item.tagName == "INPUT" || item.tagName == "textarea" || item.tagName == "TEXTAREA")) {
-			textData = item.value;
-		}
-		if (textData){
-			try {
-				// Create embedded sizeless div to contain the iframe, invisibly.
-				var div = document.getElementById('domuiprif');
-				if(div)
-					div.innerHTML = "";
-				else {
-					div = document.createElement('div');
-					div.id = 'domuiprif';
-					div.className = 'ui-printdiv';
-					document.body.appendChild(div);
-				}
-
-				//-- Create an iframe loading the required thingy.
-				var frmname = "dmuifrm"+(WebUI._frmIdCounter++);		// Create unique name to circumvent ffox "print only once" bug
-
-				$(div).html('<iframe id="'+frmname+'" name="'+frmname+'" width="1000px" height="1000px"/>'); //well, this is simple text printing, so we have some size limitations ;)
-				var frm = window.frames[frmname];
-				frm.document.open();
-				frm.document.write('<html></head></head><body style="margin:0px;">');
-				frm.document.write(textData.replace(/\n/g, '<br/>'));
-				frm.document.write('</body></html>');
-				frm.document.close();
-				frm.focus();
-				frm.print();
-			} catch(x) {
-				alert("Failed: "+x);
-			}
-		}
-	},
 
 
 	/***** Popup menu code *****/
