+++ to.etc.domui/src/main/resources/resources/ts/domui.webui.d.ts
@@ -9,6 +9,10 @@ interface WebUIStatic {
 	handleResponse(data, state): void;
 	scall(id: string, action: string, fields? : any) : void;
 
+	// busy
+	blockUI() : void;
+	unblockUI() : void;
+
 	//-- util
 	refreshElement(id: string) : void;
 	blockUI() : void;
@@ -16,6 +20,8 @@ interface WebUIStatic {
 	_T: any;
 	format(message: string, ...rest) : string;
 	getPostURL() : string;
+	getObituaryURL(): string;
+
 	_hideExpiredMessage: boolean;
 	isReallyIE7() : boolean;
 	isIE8orIE8c() : boolean;
